+++ src/sanguinehaze/charactercreator/application/UpdateNotice.java
@@ -1,4 +1,4 @@
-package sanguinehaze.charactercreator;
+package sanguinehaze.charactercreator.application;
 
 import java.awt.BorderLayout;
